refactor: MAT-269 선발 선수 리스트 컴포넌트에서 List의 뷰를 별도 컴포넌트로 추출
